requirements.txt
functions-framework==3.*
requests==2.31.0
pandas==2.0.3
python-dotenv==1.0.0
google-cloud-bigquery==3.11.4
google-cloud-storage==2.10.0
google-cloud-secret-manager
numpy==1.24.3
beautifulsoup4==4.12.3
fsspec
gcsfs
pyarrow
datetime
pytzs